rename exile command file
