+++ src/commands/exile_commands.py
@@ -1,5 +1,5 @@
 import discord
-from discord.ext.commands import Bot
+from discord.ext.commands import Bot, CommandError
 from settings import get_settings
 from services.exile_service import exile_user, unexile_user
 from util import run_command_with_logging
@@ -7,12 +7,16 @@ from util import run_command_with_logging
 
 settings = get_settings()
 
+async def is_specific_person(interaction: discord.Interaction):
+    if interaction.application_id != 1234567890:
+        raise CommandError("You do not have permission to run this command")
 
 def create_exile_commands(bot: Bot) -> None:
     @bot.tree.command()
+    @discord.app_commands.check(is_specific_person)
     @discord.app_commands.describe(user="User being exiled")
     async def unexile(interaction: discord.Interaction, user: discord.Member):
-        """Exile the specified user."""
+        """Unexile the specified user."""
 
         await run_command_with_logging(interaction, unexile_user, user)
